+++ .fvm/fvm_config.json
@@ -1,4 +1,4 @@
 {
-  "flutterSdkVersion": "3.16.5",
+  "flutterSdkVersion": "3.22.3",
   "flavors": {}
 }
